Create LICENSE.md/examplecontrol/drivingmachine
Agreement ENGLISH

APPLE INC.

TESTFLIGHT SOFTWARE LICENSE AGREEMENT TestFlight Application

PLEASE READ THIS SOFTWARE LICENSE AGREEMENT (“LICENSE”) CAREFULLY BEFORE USING THE APPLE SOFTWARE. ANY INSTALLATION AND/OR USE OF THE APPLE SOFTWARE ON THIS DEVICE OR ON YOUR OTHER SUPPORTED APPLE-BRANDED DEVICES IS SUBJECT TO THE TERMS OF THIS LICENSE, UNLESS ACCOMPANIED BY A SEPARATE LICENSE AGREEMENT. IF YOU DO NOT AGREE TO THE TERMS OF THIS LICENSE, DO NOT INSTALL AND/OR USE THE SOFTWARE ON THIS DEVICE OR ON YOUR OTHER SUPPORTED APPLE-BRANDED DEVICES. THIS LICENSE SUPERSEDES ANY PRIOR VERSIONS OF THE LICENSE INTO WHICH YOU HAVE ENTERED.

General.
A. The Apple software, and any third-party software, documentation, interfaces, fonts and any data accompanying this License whether in read only memory, on any other media or in any other form (collectively the “Apple Software”) are licensed, not sold, to you by Apple Inc. (“Apple”) for use only under the terms of this License. Apple and/or Apple’s licensors retain ownership of the Apple Software itself and reserve all rights not expressly granted to you.